third_party/WebKit/LayoutTests/http/tests/websocket/connect-error-by-no-websocket-server-closeonerror.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/connect-error-by-no-websocket-server.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description("Test that WebSocket fires error event if no WebSocket Server can be connected, and connection is closed from onerror event");

doTest(8889, true, false);

</script>
</body>
</html>
